+++ frontend/src/containers/club/clubsList/clubIcon.tsx
