+++ frontend/src/containers/event/eventsView.tsx
