+++ frontend/src/containers/event/eventsView/event.tsx
