+++ frontend/src/containers/event/eventsView/interfaces.ts
