+++ frontend/src/containers/event/eventsView/participateButton.tsx
