+++ frontend/src/containers/event/eventsView/utils.ts
