+++ frontend/src/containers/group/groupMembers/studentCard.tsx
