+++ frontend/src/containers/group/groupMembers/studentCardPicture.tsx
